Ensure e2e tests use the correct versions yaml file (#667)
We use `-ldflags` to set the `istioversion.versionsFilename` variable, but this only works when building the binary. It doesn't work when running ginkgo-based e2e tests. When the `versionsFilename` is not set via ldflags, it will now use the environment variable `VERSIONS_YAML_FILE`. If this env var isn't set, we default to `versions.yaml`.

Signed-off-by: Marko Lukša <mluksa@redhat.com>
Signed-off-by: Daniel Grimm <dgrimm@redhat.com>