Add label managed-by: sail-operator to all resources the operator deploys (#801)
We introduce a new label instead of changing the value of the `app.kubernetes.io/managed-by` label, because Helm doesn't allow us to override it. We will be able to override the label only when the operator applies the rendered manifests by itself instead of letting Helm do it. Until then, we need to resort to this additional `managed-by` label.

Signed-off-by: Marko Lukša <mluksa@redhat.com>
Signed-off-by: Daniel Grimm <dgrimm@redhat.com>