Add ConfigMap with final Helm values to charts that don't have one (#727)
* Add ConfigMap with final Helm values

Ever since the profiles were reimplemented within the Helm charts, there has been no way to see the actual values that were used to render the charts. This includes the values from the user-selected profile, the platform profile, and the compatibility version profile.

Previously, when the values from the profiles were applied before Helm was invoked, you could use `helm get values` to see the final set of values. Now, this is no longer possible, because the profile values are applied within the rendering process by the zzz_profile.yaml template and aren't stored anywhere.

This commit introduces a new ConfigMap with the name `values` for the default revision, and `values-<revision>` for all others. This ConfigMap will allow the operator (and users) to look up the actual values used to render the charts.

Signed-off-by: Marko Lukša <mluksa@redhat.com>

* Add note about `values` ConfigMap to docs

Signed-off-by: Marko Lukša <mluksa@redhat.com>

---------

Signed-off-by: Marko Lukša <mluksa@redhat.com>
Signed-off-by: Daniel Grimm <dgrimm@redhat.com>