Speed up `gen` target by only downloading charts when needed
This was driving me nuts. With this change, the `gen-charts` target runs
in 12.3s on my machine. Without it - so when downloading everything - it
takes takes 50s. On branches with fewer versions it's obviously not
quite as bad.

This will break sometimes, e.g. when making changes to the patch
functions the charts will need to be re-downloaded, which can be forced
by setting `FORCE_DOWNLOADS=true`. It is always set when executing the
`gen-check` target directly (ie in CI), to be on the safe side.

Signed-off-by: Daniel Grimm <dgrimm@redhat.com>