Remove istio versions that reached EOL
This adds an `eol: true` setting to the versions defined in
`versions.yaml`. Setting this will cause the version's charts to be
removed from the repo, rendering it uninstallable, but it will keep the
version as a valid enum value to avoid breaking API guarantees and
having to bump CRD versions.

I also added an integration test that verifies that using a version
marked as EOL results in a ReconcileError.

Signed-off-by: Daniel Grimm <dgrimm@redhat.com>