Simplify logic when using `KEEP_ON_FAILURE` (#858)
When using `KEEP_ON_FAILURE=true` also set `SKIP_CLEANUP=true`. This
keeps the cluster running after the tests, without a need to defined our
own trap logic.

The downside is that the cluster will keep running even if the tests
passed, but seeing as using the `KEEP_ON_FAILURE` flag is excpected
mainly when there are failures, it seems that it's resonable to do.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>
Signed-off-by: Daniel Grimm <dgrimm@redhat.com>