Improve e2e test cleanup by recording & cleaning up (#889)
* Consolidate namespace cleanup in a new package

The new package tracks namespaces and cleans up any namespaces it
doesn't know about.
The new `Cleaner` type can record and clean up the resources, given
it's called with `Record()` before the tests and `Cleanup()` after.

This is an initial change and more resources will be added.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* Add CRD/CR cleanup logic for tests

Now the `Cleaner` can also record and clean up any rogue CRDs/CRs that
were created during the tests.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* Replace operator uninstall with Cleanup

Now that Cleanup takes care of all resources that we care about, replace
the operator uninstall with it.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* Add docs how to use `Cleaner`

Add docs to E2E tests about the `Cleaner` and how it should be used.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

---------

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>
Signed-off-by: Daniel Grimm <dgrimm@redhat.com>