Cleaner improvements (#946)
* Print sensible error when failing to wait

Right now the error is not clear enough, adding the context of which
resource we fail to wait for should make it much easier to debug cleaner
failures.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* Make sure Record is called before cleaning up

Prevent misuse of the cleaner when trying to call the cleanup functions
without calling the `Record` function first.

Otherwise, the cleaner will attempt to remove all the resources it
tracks from the cluster.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

---------

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>
Signed-off-by: Daniel Grimm <dgrimm@redhat.com>