Deploy operator in the E2E scripts (#956)
Given that we have the `Cleaner` and each test should be cleaning after
itself, deploying the operator just once per test run will save some
test run time.
An added value is the tests are simplified and can focus on what they
need to achieve, instead of bothering with the operator setup.

Fixes: https://github.com/istio-ecosystem/sail-operator/issues/666

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>
Signed-off-by: Daniel Grimm <dgrimm@redhat.com>