Refactor integration tests (#1043)
* Use expectCondition across revision tests

Expand the method and refactor the rest of the tests to use it.
This simplifies the tests and makes them easy to read, while avoiding
the repetitiveness of condition checks.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* Extract and consolidate "no reconciliation" logic

In the integration tests weveral places were looking for controllers not
to reconcile.
Extract and place this logic in a single place so that tests become
simpler and stop repeating it.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* Consolidate expecting tag in use

This code is quite repetitive, consolidate it into a single function.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* Consolidate assertions for tag not reconciled

This code is very repetitive, consolidate it into a simple function.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* Consolidate CNI condition checking

This code is quite repetitive, consolidate it into a single function.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

---------

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>
Signed-off-by: Daniel Grimm <dgrimm@redhat.com>