Add Ambient multi cluster E2E tests for Multi Primary deployments (#1353)
* Expose common namespaces in E2E tests

To be able to use directly without redefining or passing them all over.

The namespaces are reused in the test suite files (instead of redefined)
in order to minimize the amount of changes, but we can further refactor
this in the future.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* Extract multicluster resource creation

This males tests easier to maintain while avoiding code duplication

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

* Add Ambient tests to E2E Multi Primary

Istio 1.27+ supports running Multi-Primary deployments in Abmient mode,
so tests were adapted to run both on Sidecar and Ambient mode.

Both modes run the same tests, sequentialy & independently of each
other.

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>

---------

Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>
Signed-off-by: Daniel Grimm <dgrimm@redhat.com>